requirements.txt
discord.py
aiohttp
click
process_uptime
requests
openai